+++ tests/test_configuration.py
@@ -34,13 +34,16 @@ class TestConfiguration(unittest.TestCase):
         self.assertTrue(hasattr(self.configuration, 'documentation'))
 
     def test_documentation_property_has_keys(self):
-        self.assertListEqual(['logo', 'favicon'],
+        self.assertListEqual(['logo', 'favicon', 'language'],
                              list(self.configuration.documentation.keys()))
+
+    def test_default_documentation_language_is_en(self):
+        self.assertEqual('en', self.configuration.documentation['language'])
 
     def test_has_options_property(self):
         self.assertTrue(hasattr(self.configuration, 'options'))
 
-    def test_documentation_property_has_keys(self):
+    def test_options_property_has_keys(self):
         self.assertListEqual(['logging', 'git'],
                              list(self.configuration.options.keys()))